RDM-8427 - Convert unit tests with external dependencies to IT
