+++ src/test/java/uk/gov/hmcts/ccd/data/definition/DefaultCaseDefinitionRepositoryIT.java
@@ -1,32 +1,49 @@
 package uk.gov.hmcts.ccd.data.definition;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-import javax.inject.Inject;
+import com.github.tomakehurst.wiremock.client.WireMock;
+import com.github.tomakehurst.wiremock.stubbing.StubMapping;
 import org.hamcrest.collection.IsCollectionWithSize;
+import org.junit.Assert;
 import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
 import uk.gov.hmcts.ccd.WireMockBaseTest;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseTypeDefinition;
 import uk.gov.hmcts.ccd.domain.model.definition.FieldTypeDefinition;
 import uk.gov.hmcts.ccd.domain.model.definition.JurisdictionDefinition;
 import uk.gov.hmcts.ccd.domain.model.definition.UserRole;
+import uk.gov.hmcts.ccd.endpoint.exceptions.ServiceException;
 
+import javax.inject.Inject;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static com.google.common.collect.Lists.newArrayList;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.hasItem;
-import static org.hamcrest.Matchers.hasSize;
-import static org.hamcrest.Matchers.nullValue;
+import static org.hamcrest.Matchers.*;
 import static org.hamcrest.beans.HasPropertyWithValue.hasProperty;
+import static org.hamcrest.core.StringStartsWith.startsWith;
 import static org.junit.Assert.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.ccd.domain.model.definition.FieldTypeDefinition.COLLECTION;
 import static uk.gov.hmcts.ccd.domain.model.definition.FieldTypeDefinition.COMPLEX;
 
 public class DefaultCaseDefinitionRepositoryIT extends WireMockBaseTest {
     @Inject
     private CaseDefinitionRepository caseDefinitionRepository;
+
+    private StubMapping stubMapping;
+
+    @AfterEach
+    public void tearDown() {
+        if (stubMapping != null) {
+            removeStub(stubMapping);
+        }
+    }
 
     @Test
     public void shouldGetCaseTypesForJurisdiction() {
@@ -108,5 +125,40 @@ public class DefaultCaseDefinitionRepositoryIT extends WireMockBaseTest {
             () -> assertThat(allJurisdictionDefinitions, hasItem(hasProperty("id", is("SSCS")))),
             () -> assertThat(allJurisdictionDefinitions, hasItem(hasProperty("id", is("PROBATE"))))
         );
+    }
+
+    @Test
+    public void shouldFailToGetCaseTypesForJurisdiction() {
+        stubMapping = stubFor(WireMock.get(urlMatching("/api/data/jurisdictions/server_error/case-type")).willReturn(serverError()));
+        final ServiceException exception = assertThrows(ServiceException.class,
+            () -> caseDefinitionRepository.getCaseTypesForJurisdiction("server_error"));
+        assertThat(exception.getMessage(), startsWith("Problem getting case types for the Jurisdiction:server_error because of "));
+    }
+
+    @Test
+    public void shouldFailToGetCaseType() {
+        stubMapping = stubFor(WireMock.get(urlMatching("/api/data/case-type/anything")).willReturn(serverError()));
+        final ServiceException exception = assertThrows(ServiceException.class,
+            () -> caseDefinitionRepository.getCaseType("anything"));
+        assertThat(exception.getMessage(), startsWith("Problem getting case type definition for anything because of "));
+    }
+
+    @Test
+    public void shouldFailToGetBaseTypes() {
+        when(caseDefinitionRepository.getBaseTypes()).thenCallRealMethod();
+        stubMapping = stubFor(WireMock.get(urlMatching("/api/base-types")).willReturn(serverError()));
+        final ServiceException exception = assertThrows(ServiceException.class,
+            () -> caseDefinitionRepository.getBaseTypes());
+        assertThat(exception.getMessage(), startsWith("Problem getting base types definition from definition store because of "));
+    }
+
+    @Test
+    public void shouldFailToGetClassificationsForUserRoleList() {
+        List<String> userRoles = Arrays.asList("neither_defined", "nor_defined");
+        stubMapping = stubFor(WireMock.get(urlMatching("/api/user-roles/neither_defined,nor_defined")).willReturn(serverError()));
+        final ServiceException exception = assertThrows(ServiceException.class,
+            () -> caseDefinitionRepository.getClassificationsForUserRoleList(userRoles));
+        Assert.assertThat(exception.getMessage(),
+            startsWith("Error while retrieving classification for user roles " + userRoles + " because of "));
     }
 }
